Convert all item types to intptr_t
- This is an architecture independent way to size all items the same
  size as pointers, while still retaining the ability to treat the
  items as arithmetic integers

+++ rotate.c
@@ -35,7 +35,7 @@
 #include "rotate.h"
 #include "triple-shift-rotate.h"
 
-typedef void rotate_function(int *array, size_t left, size_t right);
+typedef void rotate_function(intptr_t *array, size_t left, size_t right);
 
 typedef struct {
 	rotate_function		*rotate;
@@ -91,7 +91,7 @@ int
 main()
 {
 	struct	timespec start, end;
-	int32_t	*a;
+	intptr_t *a;
 
 	a = malloc(sizeof(*a) * MAX_VALS);
 	if (!a) {
